SSCS-7147: Timestamp of last logged into MYA in CCD UI
